- Updated `FileUploadController` to support `.xlsx` file uploads.
- Enhanced `OrderController` with logging and improved validation for customer and product selections.
- Modified `Order` model to include validation annotations and changed `TotalPrice` type to `double`.
- Updated `OrderMessage` class to reflect changes in `TotalPrice`.
- Improved `QueueService` with logging and error handling.
- Adjusted `FileUpload` model and views to restrict accepted file types and ensure proper validation.
- Updated `appsettings.json` with Azure Storage queue configuration.
Merge branch 'master' of https://github.com/VCCT-CLDV6212-2025-G1/ST10439147_CLDV6212_POE

+++ ST10439147_CLDV6212_POE/Models/OrderMessage.cs
@@ -8,7 +8,7 @@ namespace ST10439147_CLDV6212_POE.Models
         public string CustomerId { get; set; }   // Matches Customer.RowKey
         public string ProductId { get; set; }    // Matches Product.RowKey
         public int Quantity { get; set; }
-        public decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; } // Changed to double for consistency with Order model
         public DateTime OrderDate { get; set; }
 
         // Optional: Status or action type (e.g. "NewOrder", "UpdateOrder")
